# backend-api/docs/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chemistry Game Suite API",
    "description": "API cho bộ sưu tập trò chơi hóa học dựa trên OpenAPI 3.0, hỗ trợ quản lý phản ứng, ô chữ, nguyên tố, và phiên chơi. \nTìm hiểu thêm về Swagger tại [https://swagger.io](https://swagger.io).",
    "termsOfService": "https://yourchemistrygame.com/terms/",
    "contact": {
      "email": "giaphuc2004hg@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.12"
  },
...
  "servers": [
    {
      "url": "https://ct313hm01-project-gphuc04.onrender.com/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Lỗi dữ liệu đầu vào"
          },
          "error": {
            "type": "string",
            "example": "Invalid input"
          }
        },
        "required": ["code", "message"]
      },
...
    }
  },
  "paths": {
...
    "/game-sessions": {
      "get": {
        "tags": ["Game sessions"],
        "summary": "Lấy danh sách tất cả các phiên chơi",
        "description": "API để lấy danh sách tất cả các phiên chơi, yêu cầu xác thực JWT.",
        "operationId": "getAllGameSessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách phiên chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách phiên chơi thành công"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "session_id": {
                            "type": "string",
                            "example": "sess123"
                          },
                          "user_id": {
                            "type": "string",
                            "example": "user456"
                          },
                          "game_id": {
                            "type": "string",
                            "example": "game123"
                          },
                          "start_time": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-06-09T05:30:00Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Game sessions"],
        "summary": "Tạo mới một phiên chơi",
        "description": "API để tạo một phiên chơi mới, yêu cầu xác thực JWT. Hỗ trợ dữ liệu dạng multipart.",
        "operationId": "createGameSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "example": 456,
                    "description": "ID của người dùng"
                  },
                  "game_id": {
                    "type": "integer",
                    "example": 123,
                    "description": "ID của trò chơi"
                  },
                  "score": {
                    "type": "integer",
                    "example": 0,
                    "description": "Điểm số ban đầu của phiên chơi"
                  },
                  "time_spent": {
                    "type": "integer",
                    "example": 120,
                    "description": "Thời gian đã chơi (giây)"
                  },
                  "difficulty": {
                    "type": "string",
                    "enum": ["easy", "medium", "hard", "expert"],
                    "example": "easy",
                    "description": "Mức độ khó của phiên chơi"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["playing", "completed", "abandoned"],
                    "example": "playing",
                    "description": "Trạng thái của phiên chơi"
                  },
                  "game_data": {
                    "type": "string",
                    "example": "{\"level\": 1, \"progress\": 0.5}",
                    "description": "Dữ liệu trò chơi dưới dạng chuỗi JSON"
                  },
                  "started_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-06-09T05:30:00Z",
                    "description": "Thời gian bắt đầu phiên chơi"
                  },
                  "completed_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-06-09T05:35:00Z",
                    "description": "Thời gian hoàn thành phiên chơi"
                  }
                },
                "required": ["user_id", "game_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo phiên chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "POST, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo phiên chơi thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "session_id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/game-sessions/{Id}": {
      "get": {
        "tags": ["Game sessions"],
        "summary": "Lấy thông tin phiên chơi theo ID",
        "description": "API để lấy thông tin chi tiết của một phiên chơi dựa trên sessionId, yêu cầu xác thực JWT.",
        "operationId": "getGameSessionById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "ID của phiên chơi cần lấy thông tin",
            "schema": {
              "type": "string",
              "example": "sess123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin phiên chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin phiên chơi thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "session_id": {
                          "type": "string",
                          "example": "sess123"
                        },
                        "user_id": {
                          "type": "string",
                          "example": "user456"
                        },
                        "game_id": {
                          "type": "string",
                          "example": "game123"
                        },
                        "start_time": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-06-09T05:30:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID phiên chơi không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy phiên chơi",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}